trading/requirements.txt
ccxt>=4.4.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
python-telegram-bot>=21.0
schedule>=1.2.0
rich>=13.0.0
cryptography>=42.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
itsdangerous>=2.2.0
starlette>=0.41.0
